add required setting for access to map, changed markers and polygons, refactoring, fixed display issue
